started Docu (in install_readme_new). Fixed warnings for VS 64bit
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Retreat2008@4211 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/CHEMISTRY/AASequence.h
@@ -534,9 +534,9 @@ namespace OpenMS
 
 			inline ConstIterator begin() const { return ConstIterator(&peptide_, 0); }
 
-			inline Iterator end() { return Iterator(&peptide_, peptide_.size()); }
+			inline Iterator end() { return Iterator(&peptide_, (Int) peptide_.size()); }
 
-			inline ConstIterator end() const { return ConstIterator(&peptide_, peptide_.size()); }
+			inline ConstIterator end() const { return ConstIterator(&peptide_, (Int) peptide_.size()); }
 			//@}
 
 			/** @name Stream operators
